requirements.txt
fastapi[standard]
uvicorn
websockets
pyyaml
mysql-connector-python
